Very minor changes to remove compiler warnings (tested with gcc-3.4.2 under Linux)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5958 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/libmad/frame.c
@@ -315,7 +315,7 @@ int mad_header_decode(struct mad_header *header, struct mad_stream *stream)
     if (!stream->sync)
       ptr = stream->this_frame;
 
-    if (end - ptr < stream->skiplen) {
+    if (end - ptr < (long) stream->skiplen) {
       stream->skiplen   -= end - ptr;
       stream->next_frame = end;
 
@@ -400,7 +400,7 @@ int mad_header_decode(struct mad_header *header, struct mad_stream *stream)
   }
 
   /* verify there is enough data left in buffer to decode this frame */
-  if (N + MAD_BUFFER_GUARD > end - stream->this_frame) {
+  if ((long)(N + MAD_BUFFER_GUARD) > end - stream->this_frame) {
     stream->next_frame = stream->this_frame;
 
     stream->error = MAD_ERROR_BUFLEN;
