+++ apps/codecs/libmad/layer3.c
@@ -894,7 +894,7 @@ mad_fixed_t III_requantize(unsigned int value, signed int exp)
   exp += power->exponent;
 
   if (exp < 0) {
-    if (-exp >= sizeof(mad_fixed_t) * CHAR_BIT) {
+    if (-exp >= (signed int) (sizeof(mad_fixed_t) * CHAR_BIT)) {
       /* underflow */
       requantized = 0;
     }
@@ -1053,7 +1053,7 @@ enum mad_error III_huffdecode(struct mad_bitptr *ptr, mad_fixed_t xr[576],
 	  break;
 
 	case 15:
-	  if (cachesz < linbits + 2) {
+	  if ((unsigned int)cachesz < linbits + 2) {
 	    bitcache   = (bitcache << 16) | mad_bit_read(&peek, 16);
 	    cachesz   += 16;
 	    bits_left -= 16;
@@ -1088,7 +1088,7 @@ enum mad_error III_huffdecode(struct mad_bitptr *ptr, mad_fixed_t xr[576],
 	  break;
 
 	case 15:
-	  if (cachesz < linbits + 1) {
+	  if ((unsigned int)cachesz < linbits + 1) {
 	    bitcache   = (bitcache << 16) | mad_bit_read(&peek, 16);
 	    cachesz   += 16;
 	    bits_left -= 16;
